+++ client/src/app/models/network.ts
@@ -48,19 +48,16 @@ export interface NetworkProxy {
   whitelist: string[];
 }
 
-export interface NetworkNtp {
-  useNtp: boolean;
-  ntpHost: string;
-  currentTime: string;
-  timezone: string;
-  reachable: boolean;
-}
-
 export interface NetworkDateTime {
   datetime: string;
-  timezone: string;
 }
 
 export interface NetworkTimestamp {
   Timestamp: string;
+}
+
+export interface NetworkNtpReachable {
+  address: string;
+  reachable: boolean;
+  valid: boolean;
 }