+++ client/src/app/pages/settings/network/network.component.ts
@@ -6,8 +6,17 @@ import { filter } from 'rxjs/operators';
 
 import { NetworkService } from '../../../services/network.service';
 import { clone } from '../../../shared/utils';
-import { NetworkConfig, NetworkType } from '../../../models';
-import { HOST_REGEX, IP_REGEX, PORT_REGEX } from '../../../shared/utils/regex';
+import {
+  NetworkConfig,
+  NetworkNtpReachable,
+  NetworkType
+} from '../../../models';
+import {
+  HOST_REGEX,
+  IP_REGEX,
+  NETMASK_REGEX,
+  PORT_REGEX
+} from '../../../shared/utils/regex';
 import { Status } from 'app/shared/state';
 
 @Component({
@@ -20,10 +29,12 @@ export class NetworkComponent implements OnInit, OnDestroy {
 
   config!: NetworkConfig;
   originalConfig!: NetworkConfig;
+  ntpReachable!: NetworkNtpReachable[];
 
   hostRegex = HOST_REGEX;
   portRegex = PORT_REGEX;
   ipRegex = IP_REGEX;
+  netmaskRegex = NETMASK_REGEX;
 
   get showLoading() {
     return (
@@ -32,21 +43,8 @@ export class NetworkComponent implements OnInit, OnDestroy {
     );
   }
 
-  get timezoneOptionsSearchResults() {
-    if (!this.timezoneOptions || !this.timezoneOptionKeyphrase) {
-      return this.timezoneOptions;
-    }
-    return this.timezoneOptions.filter((x) =>
-      x
-        .toLocaleLowerCase()
-        .includes(this.timezoneOptionKeyphrase.toLocaleLowerCase())
-    );
-  }
-
   sub: Subscription = new Subscription();
   selectedTab!: string;
-  timezoneOptions!: string[];
-  timezoneOptionKeyphrase = '';
 
   @ViewChild('mainForm') mainForm!: NgForm;
 
@@ -58,15 +56,19 @@ export class NetworkComponent implements OnInit, OnDestroy {
         .pipe(filter((el) => !!el))
         .subscribe((x) => this.onConfig(x))
     );
-
-    this.timezoneOptions = this._getTimezoneOptions();
+    this.sub.add(
+      this.networkService.ntpReachable
+        .pipe(filter((el) => !!el))
+        .subscribe((x) => this.onNtpReachable(x))
+    );
 
     this.sub.add(this.networkService.setNetworkAdaptersTimer().subscribe());
 
-    this.networkService.getNetworkAdapters();
-    this.networkService.getNetworkProxy();
-    this.networkService.getNetworkNtp();
-    this.networkService.getNetworkTimestamp();
+    this.networkService
+      .getNetworkAdapters()
+      .then(() => this.networkService.getNetworkProxy())
+      .then(() => this.networkService.getNetworkNtp())
+      .then(() => this.networkService.getNetworkTimestamp());
   }
 
   get tabs() {
@@ -85,9 +87,19 @@ export class NetworkComponent implements OnInit, OnDestroy {
     this.config = clone(x);
     this.originalConfig = clone(x);
 
-    if (!this.selectedTab) {
+    if (!this.tabs.includes(this.selectedTab)) {
       this.selectedTab = this.tabs[0];
     }
+  }
+
+  private onNtpReachable(x: NetworkNtpReachable[]) {
+    this.ntpReachable = x;
+  }
+
+  getNtpReachableByAddress(address: string): boolean {
+    return this.ntpReachable.find(
+      (obj: NetworkNtpReachable) => obj.address === address
+    )?.reachable;
   }
 
   onSelectTab(tab: string) {
@@ -121,24 +133,5 @@ export class NetworkComponent implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.sub.unsubscribe();
-  }
-
-  private _getTimezoneOptions() {
-    return moment.tz
-      .names()
-      .filter((name) =>
-        [
-          'Universal',
-          'Africa/',
-          'America/',
-          'Antarctica/',
-          'Arctic/',
-          'Asia/',
-          'Australia/',
-          'Europe/',
-          'Indian/',
-          'Pacific/'
-        ].find((x) => name.startsWith(x))
-      );
   }
 }