+++ client/src/app/services/network.service.ts
@@ -10,102 +10,20 @@ import { errorHandler } from '../shared/utils';
 import {
   NetworkAdapter,
   NetworkConfig,
-  NetworkProxy,
   NetworkDateTime,
-  NetworkTimestamp
+  NetworkProxy,
+  NetworkTimestamp,
+  NetworkNtpReachable
 } from '../models';
-import { ConfigurationAgentHttpMockupService } from 'app/shared';
+import { ConfigurationAgentHttpService, HttpService } from 'app/shared';
 export interface NetworkState {
   status: Status;
   adapters: NetworkAdapter[];
   proxy: NetworkProxy;
   ntp: string[];
   timestamp: NetworkDateTime;
+  ntpReachable: NetworkNtpReachable[];
 }
-
-// TODO: Connect to Network API
-let RESPONSE_ADAPTERS: NetworkAdapter[] = [
-  {
-    id: 'x1',
-    displayName: 'Ethernet X1 p1',
-    enabled: false,
-    ipv4Settings: {
-      enabled: false,
-      dhcp: true,
-      ipAddresses: [
-        {
-          Address: '',
-          Netmask: ''
-        }
-      ],
-      defaultGateway: '',
-      dnsserver: ['192.168.214.230', '192.168.214.230']
-    },
-    ipv6Settings: {
-      enabled: false,
-      dhcp: false,
-      ipAddresses: [
-        {
-          Address: '',
-          Netmask: ''
-        }
-      ],
-      defaultGateway: '',
-      dnsserver: []
-    },
-    macAddress: '',
-    ssid: ''
-  },
-  {
-    id: 'x2',
-    displayName: 'Ethernet X2 p1',
-    enabled: false,
-    ipv4Settings: {
-      enabled: false,
-      dhcp: true,
-      ipAddresses: [
-        {
-          Address: '192.168.214.230',
-          Netmask: '255.255.255.0'
-        }
-      ],
-      defaultGateway: '192.168',
-      dnsserver: ['192.168.214.230', '192.168.214.230']
-    },
-    ipv6Settings: {
-      enabled: false,
-      dhcp: false,
-      ipAddresses: [
-        {
-          Address: '',
-          Netmask: ''
-        }
-      ],
-      defaultGateway: '',
-      dnsserver: []
-    },
-    macAddress: '',
-    ssid: ''
-  }
-] as any;
-
-// TODO: Connect to Network API
-let RESPONSE_PROXY: NetworkProxy = {
-  enabled: true,
-  host: '',
-  port: 62145,
-  username: '',
-  password: '',
-  whitelist: ['']
-} as any;
-
-// TODO: Connect to Network API
-let RESPONSE_NTP: string[] = ['time.dmgmori.net'] as any;
-
-// TODO: Connect to Network API
-let RESPONSE_TIMESTAMP: NetworkTimestamp = {
-  Timestamp: 'Sun Jan 09 2022 00:14:58 GMT+0000 (Coordinated Universal Time)'
-} as any;
 
 @Injectable()
 export class NetworkService {
@@ -113,7 +31,8 @@ export class NetworkService {
 
   constructor(
     storeFactory: StoreFactory<NetworkState>,
-    private configurationAgentHttpMockupService: ConfigurationAgentHttpMockupService,
+    private configurationAgentHttpService: ConfigurationAgentHttpService,
+    private httpService: HttpService,
     private translate: TranslateService,
     private toastr: ToastrService
   ) {
@@ -154,6 +73,13 @@ export class NetworkService {
     );
   }
 
+  get ntpReachable() {
+    return this._store.state.pipe(
+      filter((x) => x.status != Status.NotInitialized),
+      map((x) => x.ntpReachable)
+    );
+  }
+
   setNetworkAdaptersTimer() {
     return interval(10 * 1000).pipe(
       mergeMap(() => from(this.getNetworkAdapters()))
@@ -162,12 +88,15 @@ export class NetworkService {
 
   async getNetworkAdapters() {
     try {
-      const response = await this.configurationAgentHttpMockupService.get<
+      const response = await this.configurationAgentHttpService.get<
         NetworkAdapter[]
-      >(`/network/adapters`, undefined, RESPONSE_ADAPTERS);
+      >(`/network/adapters`);
+
+      const verifiedAdapters = this._deserializeNetworkAdapters(response);
+
       this._store.patchState((state) => {
         state.status = Status.Ready;
-        state.adapters = response;
+        state.adapters = verifiedAdapters;
       });
     } catch (err) {
       this.toastr.error(this.translate.instant('settings-network.LoadError'));
@@ -186,15 +115,18 @@ export class NetworkService {
     const verifiedObj = this._serializeNetworkAdapters(obj);
 
     try {
-      this.setMockDataById(verifiedObj);
       const response =
-        await this.configurationAgentHttpMockupService.put<NetworkAdapter>(
+        await this.configurationAgentHttpService.put<NetworkAdapter>(
           `/network/adapters/${obj.id}`,
           verifiedObj
         );
       this._store.patchState((state) => {
         state.status = Status.Ready;
-        state.adapters = RESPONSE_ADAPTERS;
+        Object.keys(state.adapters).forEach((key) => {
+          if (state.adapters[key].id.includes(obj.id)) {
+            state.adapters[key] = response;
+          }
+        });
       });
     } catch (err) {
       this.toastr.error(this.translate.instant('settings-network.UpdateError'));
@@ -208,10 +140,8 @@ export class NetworkService {
   async getNetworkProxy() {
     try {
       const response =
-        await this.configurationAgentHttpMockupService.get<NetworkProxy>(
-          `/network/proxy`,
-          undefined,
-          RESPONSE_PROXY
+        await this.configurationAgentHttpService.get<NetworkProxy>(
+          `/network/proxy`
         );
       this._store.patchState((state) => {
         state.status = Status.Ready;
@@ -232,7 +162,7 @@ export class NetworkService {
     });
 
     try {
-      await this.configurationAgentHttpMockupService.put<NetworkProxy>(
+      await this.configurationAgentHttpService.put<NetworkProxy>(
         `/network/proxy`,
         obj
       );
@@ -251,13 +181,14 @@ export class NetworkService {
 
   async getNetworkNtp() {
     try {
-      const response = await this.configurationAgentHttpMockupService.get<
-        string[]
-      >(`/network/ntp`, undefined, RESPONSE_NTP);
+      const response = await this.configurationAgentHttpService.get<string[]>(
+        `/network/ntp`
+      );
       this._store.patchState((state) => {
         state.status = Status.Ready;
         state.ntp = response;
       });
+      !!response?.[0] && this.getNtpReachable(response);
     } catch (err) {
       this.toastr.error(this.translate.instant('settings-network.LoadError'));
       errorHandler(err);
@@ -270,10 +201,11 @@ export class NetworkService {
   async updateNetworkNtp(obj: string[]) {
     this._store.patchState((state) => {
       state.status = Status.Loading;
+      state.ntpReachable = [{ address: '', reachable: false, valid: false }];
     });
 
     try {
-      await this.configurationAgentHttpMockupService.put<string[]>(
+      await this.configurationAgentHttpService.put<string[]>(
         `/network/ntp`,
         obj
       );
@@ -281,8 +213,28 @@ export class NetworkService {
         state.status = Status.Ready;
         state.ntp = obj;
       });
+      !!obj?.[0] && this.getNtpReachable(obj);
     } catch (err) {
       this.toastr.error(this.translate.instant('settings-network.UpdateError'));
+      errorHandler(err);
+      this._store.patchState(() => ({
+        status: Status.Failed
+      }));
+    }
+  }
+
+  async getNtpReachable(ntp: string[]) {
+    try {
+      const response = await this.httpService.get<NetworkNtpReachable[]>(
+        `/check-ntp?`,
+        { params: { addresses: ntp }, responseType: 'json' }
+      );
+      this._store.patchState((state) => {
+        state.status = Status.Ready;
+        state.ntpReachable = response;
+      });
+    } catch (err) {
+      this.toastr.error(this.translate.instant('settings-network.LoadError'));
       errorHandler(err);
       this._store.patchState(() => ({
         status: Status.Failed
@@ -293,10 +245,8 @@ export class NetworkService {
   async getNetworkTimestamp() {
     try {
       const response =
-        await this.configurationAgentHttpMockupService.get<NetworkTimestamp>(
-          `/system/time`,
-          undefined,
-          RESPONSE_TIMESTAMP
+        await this.configurationAgentHttpService.get<NetworkTimestamp>(
+          `/system/time`
         );
 
       const verifiedObj = this._serializeNetworkTimestamp(response);
@@ -322,7 +272,7 @@ export class NetworkService {
     const verifiedObj = this._deserializeNetworkTimestamp(obj);
 
     try {
-      await this.configurationAgentHttpMockupService.put<NetworkTimestamp>(
+      await this.configurationAgentHttpService.put<NetworkTimestamp>(
         `/system/time`,
         verifiedObj
       );
@@ -339,17 +289,10 @@ export class NetworkService {
     }
   }
 
-  setMockDataById(obj: NetworkAdapter) {
-    Object.keys(RESPONSE_ADAPTERS).forEach((key) => {
-      if (RESPONSE_ADAPTERS[key].id.includes(obj.id)) {
-        RESPONSE_ADAPTERS[key] = obj;
-      }
-    });
-  }
-
   private _emptyState() {
     return <NetworkState>{
-      status: Status.NotInitialized
+      status: Status.NotInitialized,
+      ntpReachable: [{ address: '', reachable: false, valid: false }]
     };
   }
 
@@ -368,17 +311,58 @@ export class NetworkService {
         };
       }
     }
-    return adapter;
+    const maskNodes: string[] =
+      adapter.ipv4Settings.ipAddresses[0].Netmask.match(/(\d+)/g);
+    let cidr = 0;
+    for (let i in maskNodes) {
+      cidr += ((+maskNodes[i] >>> 0).toString(2).match(/1/g) || []).length;
+    }
+    return {
+      ...adapter,
+      ipv4Settings: {
+        ...adapter.ipv4Settings,
+        ipAddresses: [
+          {
+            Address: adapter.ipv4Settings.ipAddresses[0].Address,
+            Netmask: String(cidr)
+          }
+        ]
+      }
+    };
+  }
+
+  private _deserializeNetmask(netmask: unknown): number[] {
+    const mask = 0xffffffff << (32 - Number(netmask));
+    return [mask >>> 24, (mask >> 16) & 0xff, (mask >> 8) & 0xff, mask & 0xff];
+  }
+
+  private _deserializeNetworkAdapters(
+    adapters: NetworkAdapter[]
+  ): NetworkAdapter[] {
+    return adapters.map((adapter) => {
+      const address = adapter?.ipv4Settings?.ipAddresses[0];
+      return {
+        ...adapter,
+        ipv4Settings: {
+          ...adapter.ipv4Settings,
+          ipAddresses: [
+            {
+              Address: address?.Address,
+              Netmask:
+                address?.Netmask &&
+                this._deserializeNetmask(address.Netmask).join('.')
+            }
+          ]
+        }
+      };
+    });
   }
 
   private _serializeNetworkTimestamp(
     timestamp: NetworkTimestamp
   ): NetworkDateTime {
     return {
-      datetime: moment(timestamp?.Timestamp)
-        .parseZone()
-        .format('YYYY-MM-DDTHH:mm:ss'),
-      timezone: 'Universal'
+      datetime: moment(timestamp?.Timestamp).format('YYYY-MM-DDTHH:mm:ss')
     };
   }
 
@@ -387,8 +371,9 @@ export class NetworkService {
   ): NetworkTimestamp {
     return {
       Timestamp: moment
-        .tz(timestamp.datetime, timestamp.timezone)
-        .toISOString(true)
+        .utc(timestamp.datetime)
+        .add(moment(timestamp.datetime).utcOffset(), 'm')
+        .toISOString()
     };
   }
 }