+++ client/src/app/services/system-information.service.ts
@@ -4,7 +4,7 @@ import { ToastrService } from 'ngx-toastr';
 import { TranslateService } from '@ngx-translate/core';
 
 import {
-  ConfigurationAgentHttpMockupService,
+  ConfigurationAgentHttpService,
   HttpService,
   RequestOptionsArgs
 } from '../shared';
@@ -39,7 +39,7 @@ export class SystemInformationService {
   constructor(
     storeFactory: StoreFactory<SystemInformationState>,
     private httpService: HttpService,
-    private configurationAgentHttpMockupService: ConfigurationAgentHttpMockupService,
+    private configurationAgentHttpService: ConfigurationAgentHttpService,
     private translate: TranslateService,
     private toastr: ToastrService
   ) {
@@ -96,7 +96,7 @@ export class SystemInformationService {
   }
 
   async getServerTime(): Promise<string> {
-    const response = await this.configurationAgentHttpMockupService.get<{
+    const response = await this.configurationAgentHttpService.get<{
       Timestamp: string;
     }>(`/system/time`);
 
@@ -105,14 +105,21 @@ export class SystemInformationService {
 
   async restartDevice(): Promise<boolean> {
     try {
-      await this.configurationAgentHttpMockupService.post(
-        `/system/restart`,
-        null
-      );
+      await this.configurationAgentHttpService.post(`/system/restart`, null);
 
       return true;
     } catch (e) {
       return true;
+    }
+  }
+
+  async factoryReset(): Promise<boolean> {
+    try {
+      await this.httpService.post(`/systemInfo/factoryreset`, null);
+
+      return true;
+    } catch (e) {
+      return false;
     }
   }